refactor(mixer_module): change MixingOutput to use float outputs
MixingOutput now passes float values to output drivers instead of
uint16_t. This removes the need for the 8192 offset encoding and
allows reversible motors to receive negative values directly.

+++ src/drivers/actuators/vertiq_io/vertiq_io.hpp
@@ -34,6 +34,7 @@
 
 #include <drivers/device/device.h>
 #include <lib/led/led.h>
+#include <lib/mathlib/mathlib.h>
 #include <lib/mixer_module/mixer_module.hpp>
 #include <lib/perf/perf_counter.h>
 
@@ -94,14 +95,14 @@ public:
 	void print_info();
 
 	/** @see OutputModuleInterface */
-	bool updateOutputs(uint16_t outputs[MAX_ACTUATORS],
+	bool updateOutputs(float outputs[MAX_ACTUATORS],
 			   unsigned num_outputs, unsigned num_control_groups_updated) override;
 
 	/**
 	* @brief Used to package and transmit controls via IQUART
 	* @param outputs The output throttles calculated by the mixer
 	*/
-	void OutputControls(uint16_t outputs[MAX_ACTUATORS]);
+	void OutputControls(float outputs[MAX_ACTUATORS]);
 
 private:
